+++ src/charade/helpers/get-charade-field-content.helper.ts
@@ -1,0 +1,6 @@
+export const getCharadeFieldContent = (
+  contents: (string | undefined)[],
+  index: number,
+  rowIndex: number,
+  count: number,
+) => contents[index + (rowIndex * (count + 1))] ?? ''